+++ src/main/java/calculations/Program.java
@@ -1,14 +1,16 @@
-package program;
+package calculations;
 
-import org.springframework.boot.SpringApplication;
+import calculations.controller.CalculationController;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-@SpringBootApplication()
+@SpringBootApplication
 public class Program {
     public static void main(String[] args) {
 //        SpringApplication.run(Program.class, args);
+
         ApplicationContext context = new ClassPathXmlApplicationContext("beans.xml");
+        context.getBean(CalculationController.class).start();
     }
 }
