+++ src/main/java/calculations/model/calculator/AvgCalculator.java
@@ -1,5 +1,8 @@
-package model.calculator;
+package calculations.model.calculator;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class AvgCalculator extends AbstractCalculation {
 
     public AvgCalculator(int[] input) {
