+++ src/main/java/calculations/model/calculator/Calculation.java
@@ -1,4 +1,4 @@
-package model.calculator;
+package calculations.model.calculator;
 
 public interface Calculation {
     int execute();
