+++ src/main/java/calculations/model/calculator/MinCalculator.java
@@ -1,5 +1,8 @@
-package model.calculator;
+package calculations.model.calculator;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class MinCalculator extends  AbstractCalculation{
     public MinCalculator(int[] input) {
         super(input, "Min");
